structure js files
