+++ dev/index.html
@@ -32,7 +32,7 @@
       </section>
 
       <!-- PILAR -->
-      <div ng-controller="mainController">
+      <div ng-controller="calenderController">
         <div class="arrow-container">
           <div class="left-arrow"><i class="flaticon-arrow487" ng-click="minusWeek()"></i></div>
           <div class="week-nr">Rubrik</div>
