--- dev/js/app/script.js
@@ -1,6 +1,0 @@
-/**
- * Created by Vilhelm on 15-11-30.
- */
-$( document ).ready(function() {
-// GO!
-});
