+++ dev/js/main.js
@@ -1,188 +1,200 @@
-var app = angular.module('booker', []);
+(function() {
+  'use strict';
 
-app.controller('mainController', ['$scope', function($scope) {
-  var milliSeconds = 24 * 60 * 60 * 1000;
-  $scope.daySwitch = 0;
+  angular.module('booker', []);
 
-  function createCalender()
-  {
-    $scope.date = new Date(new Date().getTime() + $scope.daySwitch*milliSeconds);
-    $scope.dates = [];
+})();
+
+(function() {
+  'use strict';
+
+  angular.module("ngLocale", [], ["$provide", function($provide) {
+    var PLURAL_CATEGORY = {ZERO: "zero", ONE: "one", TWO: "two", FEW: "few", MANY: "many", OTHER: "other"};
+    function getDecimals(n) {
+      n = n + '';
+      var i = n.indexOf('.');
+      return (i == -1) ? 0 : n.length - i - 1;
+    }
+
+    function getVF(n, opt_precision) {
+      var v = opt_precision;
+
+      if (undefined === v) {
+        v = Math.min(getDecimals(n), 3);
+      }
+
+      var base = Math.pow(10, v);
+      var f = ((n * base) | 0) % base;
+      return {v: v, f: f};
+    }
+
+    $provide.value("$locale", {
+      "DATETIME_FORMATS": {
+        "AMPMS": [
+          "fm",
+          "em"
+        ],
+        "DAY": [
+          "s\u00f6ndag",
+          "m\u00e5ndag",
+          "tisdag",
+          "onsdag",
+          "torsdag",
+          "fredag",
+          "l\u00f6rdag"
+        ],
+        "ERANAMES": [
+          "f\u00f6re Kristus",
+          "efter Kristus"
+        ],
+        "ERAS": [
+          "f.Kr.",
+          "e.Kr."
+        ],
+        "FIRSTDAYOFWEEK": 0,
+        "MONTH": [
+          "januari",
+          "februari",
+          "mars",
+          "april",
+          "maj",
+          "juni",
+          "juli",
+          "augusti",
+          "september",
+          "oktober",
+          "november",
+          "december"
+        ],
+        "SHORTDAY": [
+          "s\u00f6n",
+          "m\u00e5n",
+          "tis",
+          "ons",
+          "tors",
+          "fre",
+          "l\u00f6r"
+        ],
+        "SHORTMONTH": [
+          "jan.",
+          "feb.",
+          "mars",
+          "apr.",
+          "maj",
+          "juni",
+          "juli",
+          "aug.",
+          "sep.",
+          "okt.",
+          "nov.",
+          "dec."
+        ],
+        "STANDALONEMONTH": [
+          "Januari",
+          "Februari",
+          "Mars",
+          "April",
+          "Maj",
+          "Juni",
+          "Juli",
+          "Augusti",
+          "September",
+          "Oktober",
+          "November",
+          "December"
+        ],
+        "WEEKENDRANGE": [
+          5,
+          6
+        ],
+        "fullDate": "EEEE d MMMM y",
+        "longDate": "d MMMM y",
+        "medium": "d MMM y HH:mm:ss",
+        "mediumDate": "d MMM y",
+        "mediumTime": "HH:mm:ss",
+        "short": "y-MM-dd HH:mm",
+        "shortDate": "y-MM-dd",
+        "shortTime": "HH:mm"
+      },
+      "NUMBER_FORMATS": {
+        "CURRENCY_SYM": "kr",
+        "DECIMAL_SEP": ",",
+        "GROUP_SEP": "\u00a0",
+        "PATTERNS": [
+          {
+            "gSize": 3,
+            "lgSize": 3,
+            "maxFrac": 3,
+            "minFrac": 0,
+            "minInt": 1,
+            "negPre": "-",
+            "negSuf": "",
+            "posPre": "",
+            "posSuf": ""
+          },
+          {
+            "gSize": 3,
+            "lgSize": 3,
+            "maxFrac": 2,
+            "minFrac": 2,
+            "minInt": 1,
+            "negPre": "-",
+            "negSuf": "\u00a0\u00a4",
+            "posPre": "",
+            "posSuf": "\u00a0\u00a4"
+          }
+        ]
+      },
+      "id": "sv-se",
+      "pluralCat": function(n, opt_precision) {  var i = n | 0;  var vf = getVF(n, opt_precision);  if (i == 1 && vf.v == 0) {    return PLURAL_CATEGORY.ONE;  }  return PLURAL_CATEGORY.OTHER;}
+    });
+  }]);
+
+})();
+
+(function() {
+  'use strict';
+
+  angular.module('booker')
+    .controller('calenderController', ['$scope', 'calenderController']);
+
+  function calenderController($scope) {
+
+    var milliSeconds = 24 * 60 * 60 * 1000;
+    $scope.daySwitch = 0;
+
+    function createCalender()
+    {
+      $scope.date = new Date(new Date().getTime() + $scope.daySwitch*milliSeconds);
+      $scope.dates = [];
       var d;
       for(var i = 0; i < 7; i++)
       {
-          var day = {};
-          d = $scope.date;
-          d = (d.setDate((d.getDate() + 1)));
-          day.dayName = d;
-          $scope.dates.push(day);
+        var day = {};
+        d = $scope.date;
+        d = (d.setDate((d.getDate() + 1)));
+        day.dayName = d;
+        $scope.dates.push(day);
       }
-  }
-  createCalender()
+    }
+    createCalender();
     $scope.plusWeek = function() {
 
       if($scope.daySwitch < 14)
       {
         $scope.daySwitch += 7;
-        createCalender()
+        createCalender();
       }
-  	};
+    };
 
-   	 $scope.minusWeek = function() {
-       if($scope.daySwitch > 0)
-       {
-         $scope.daySwitch -= 7;
-         createCalender()
-       }
-  	};
-}]);
-
-/**
- * Created by Vilhelm on 15-11-30.
- */
-$( document ).ready(function() {
-// GO!
-});
-
-'use strict';
-angular.module("ngLocale", [], ["$provide", function($provide) {
-var PLURAL_CATEGORY = {ZERO: "zero", ONE: "one", TWO: "two", FEW: "few", MANY: "many", OTHER: "other"};
-function getDecimals(n) {
-  n = n + '';
-  var i = n.indexOf('.');
-  return (i == -1) ? 0 : n.length - i - 1;
-}
-
-function getVF(n, opt_precision) {
-  var v = opt_precision;
-
-  if (undefined === v) {
-    v = Math.min(getDecimals(n), 3);
+    $scope.minusWeek = function() {
+      if($scope.daySwitch > 0)
+      {
+        $scope.daySwitch -= 7;
+        createCalender();
+      }
+    };
   }
 
-  var base = Math.pow(10, v);
-  var f = ((n * base) | 0) % base;
-  return {v: v, f: f};
-}
 
-$provide.value("$locale", {
-  "DATETIME_FORMATS": {
-    "AMPMS": [
-      "fm",
-      "em"
-    ],
-    "DAY": [
-      "s\u00f6ndag",
-      "m\u00e5ndag",
-      "tisdag",
-      "onsdag",
-      "torsdag",
-      "fredag",
-      "l\u00f6rdag"
-    ],
-    "ERANAMES": [
-      "f\u00f6re Kristus",
-      "efter Kristus"
-    ],
-    "ERAS": [
-      "f.Kr.",
-      "e.Kr."
-    ],
-    "FIRSTDAYOFWEEK": 0,
-    "MONTH": [
-      "januari",
-      "februari",
-      "mars",
-      "april",
-      "maj",
-      "juni",
-      "juli",
-      "augusti",
-      "september",
-      "oktober",
-      "november",
-      "december"
-    ],
-    "SHORTDAY": [
-      "s\u00f6n",
-      "m\u00e5n",
-      "tis",
-      "ons",
-      "tors",
-      "fre",
-      "l\u00f6r"
-    ],
-    "SHORTMONTH": [
-      "jan.",
-      "feb.",
-      "mars",
-      "apr.",
-      "maj",
-      "juni",
-      "juli",
-      "aug.",
-      "sep.",
-      "okt.",
-      "nov.",
-      "dec."
-    ],
-    "STANDALONEMONTH": [
-      "Januari",
-      "Februari",
-      "Mars",
-      "April",
-      "Maj",
-      "Juni",
-      "Juli",
-      "Augusti",
-      "September",
-      "Oktober",
-      "November",
-      "December"
-    ],
-    "WEEKENDRANGE": [
-      5,
-      6
-    ],
-    "fullDate": "EEEE d MMMM y",
-    "longDate": "d MMMM y",
-    "medium": "d MMM y HH:mm:ss",
-    "mediumDate": "d MMM y",
-    "mediumTime": "HH:mm:ss",
-    "short": "y-MM-dd HH:mm",
-    "shortDate": "y-MM-dd",
-    "shortTime": "HH:mm"
-  },
-  "NUMBER_FORMATS": {
-    "CURRENCY_SYM": "kr",
-    "DECIMAL_SEP": ",",
-    "GROUP_SEP": "\u00a0",
-    "PATTERNS": [
-      {
-        "gSize": 3,
-        "lgSize": 3,
-        "maxFrac": 3,
-        "minFrac": 0,
-        "minInt": 1,
-        "negPre": "-",
-        "negSuf": "",
-        "posPre": "",
-        "posSuf": ""
-      },
-      {
-        "gSize": 3,
-        "lgSize": 3,
-        "maxFrac": 2,
-        "minFrac": 2,
-        "minInt": 1,
-        "negPre": "-",
-        "negSuf": "\u00a0\u00a4",
-        "posPre": "",
-        "posSuf": "\u00a0\u00a4"
-      }
-    ]
-  },
-  "id": "sv-se",
-  "pluralCat": function(n, opt_precision) {  var i = n | 0;  var vf = getVF(n, opt_precision);  if (i == 1 && vf.v == 0) {    return PLURAL_CATEGORY.ONE;  }  return PLURAL_CATEGORY.OTHER;}
-});
-}]);
+})();
